Retina Mapping Project/requirements.txt
absl-py
affine
alphashape
ansicon
astunparse
attrs
autopep8
blessed
bottle
bottle-websocket
brotlipy
cachetools
certifi
cffi
charset-normalizer
click
click-log
click-plugins
cligj
colorama
commonmark
conda
contourpy
cycler
decorator
dill
distlib
Eel
filelock
Flask
flatbuffers
fonttools
future
gast
gevent
gevent-websocket
google-auth
google-auth-oauthlib
google-pasta
greenlet
grpcio
h5py
idna
imageio
imageio-ffmpeg
importlib-metadata
itsdangerous
Jinja2
jinxed
joblib
keras
Keras-Preprocessing
kiwisolver
libclang
llvmlite
Markdown
MarkupSafe
matplotlib
menuinst
moviepy
networkx
npyscreen
numba
numpy
oauthlib
objexplore
opencv-python
opt-einsum
packaging
pandas
Pillow
platformdirs
proglog
protobuf
pyasn1
pyasn1-modules
pyassimp
pycodestyle
pycosat
pycparser
Pygments
pyopencl
pyparsing
python-dateutil
pytools
pytz
PyWavelets
pywin32
rasterio
requests
requests-oauthlib
rich
rsa
Rtree
scikit-image
scikit-learn
scipy
Shapely
snuggs
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-cpu
tensorflow-directml-plugin
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tifffile
toml
trimesh
typing_extensions
urllib3
virtualenv
wcwidth
Werkzeug
whichcraft
wincertstore
windows-curses
wrapt
wxPython
zipp
zope.event
zope.interface